+++ backend/src/ee/services/permission/default-roles.ts
@@ -7,7 +7,6 @@ import {
   ProjectPermissionApprovalRequestActions,
   ProjectPermissionApprovalRequestGrantActions,
   ProjectPermissionAuditLogsActions,
-  ProjectPermissionAuditReportActions,
   ProjectPermissionCertificateActions,
   ProjectPermissionCertificateAuthorityActions,
   ProjectPermissionCertificatePolicyActions,
@@ -497,15 +496,13 @@ const buildAdminPermissionRules = () => {
 
   can([ProjectPermissionSecretApprovalRequestActions.Read], ProjectPermissionSub.SecretApprovalRequest);
 
-  can([ProjectPermissionInsightsActions.Read], ProjectPermissionSub.Insights);
-
   can(
     [
-      ProjectPermissionAuditReportActions.Create,
-      ProjectPermissionAuditReportActions.Read,
-      ProjectPermissionAuditReportActions.Delete
+      ProjectPermissionInsightsActions.Read,
+      ProjectPermissionInsightsActions.GenerateReport,
+      ProjectPermissionInsightsActions.DeleteReport
     ],
-    ProjectPermissionSub.AuditReports
+    ProjectPermissionSub.Insights
   );
 
   return rules;
@@ -642,7 +639,7 @@ const buildMemberPermissionRules = () => {
 
   can([ProjectPermissionActions.Read], ProjectPermissionSub.Role);
   can([ProjectPermissionAuditLogsActions.Read], ProjectPermissionSub.AuditLogs);
-  can([ProjectPermissionAuditReportActions.Read], ProjectPermissionSub.AuditReports);
+  can([ProjectPermissionInsightsActions.Read], ProjectPermissionSub.Insights);
   can([ProjectPermissionActions.Read], ProjectPermissionSub.IpAllowList);
 
   can([ProjectPermissionCertificateAuthorityActions.Read], ProjectPermissionSub.CertificateAuthorities);
@@ -773,7 +770,7 @@ const buildViewerPermissionRules = () => {
   can(ProjectPermissionActions.Read, ProjectPermissionSub.Environments);
   can(ProjectPermissionActions.Read, ProjectPermissionSub.Tags);
   can(ProjectPermissionAuditLogsActions.Read, ProjectPermissionSub.AuditLogs);
-  can(ProjectPermissionAuditReportActions.Read, ProjectPermissionSub.AuditReports);
+  can(ProjectPermissionInsightsActions.Read, ProjectPermissionSub.Insights);
   can(ProjectPermissionActions.Read, ProjectPermissionSub.IpAllowList);
   can(ProjectPermissionCertificateAuthorityActions.Read, ProjectPermissionSub.CertificateAuthorities);
   can(ProjectPermissionCertificateActions.Read, ProjectPermissionSub.Certificates);
