+++ backend/src/services/smtp/emails/AuditReportTemplate.tsx
@@ -10,13 +10,13 @@ interface AuditReportTemplateProps extends Omit<BaseEmailWrapperProps, "title" |
 
 export const AuditReportTemplate = ({ siteUrl, projectName, reports }: AuditReportTemplateProps) => {
   return (
-    <BaseEmailWrapper title="Audit Report" preview="Your requested audit report is ready." siteUrl={siteUrl}>
+    <BaseEmailWrapper title="Report" preview="Your requested report is ready." siteUrl={siteUrl}>
       <Heading className="text-black text-[18px] leading-[28px] text-center font-normal p-0 mx-0">
-        Your <strong>audit report</strong> is ready
+        Your <strong>report</strong> is ready
       </Heading>
       <Section className="px-[24px] mt-[36px] pt-[8px] pb-[8px] text-[14px] border border-solid border-gray-200 rounded-md bg-gray-50">
         <Text className="text-[14px]">
-          The audit report you requested for the project <strong>{projectName}</strong> has been generated. The
+          The report you requested for the project <strong>{projectName}</strong> has been generated. The
           following {reports.length === 1 ? "report is" : "reports are"} attached to this email
           {reports.length === 1 ? " as a CSV file" : ", each as its own CSV file"}:
         </Text>
