+++ frontend/src/hooks/api/auditReports/mutations.tsx
@@ -9,7 +9,7 @@ export const useRequestAuditReport = () => {
   const queryClient = useQueryClient();
   return useMutation({
     mutationFn: async (dto: TRequestAuditReportDTO) => {
-      const { data } = await apiRequest.post<TAuditReport>("/api/v1/audit-reports", dto);
+      const { data } = await apiRequest.post<TAuditReport>("/api/v1/insights/secrets/reports", dto);
       return data;
     },
     onSuccess: () => queryClient.invalidateQueries({ queryKey: auditReportKeys.all() })
@@ -21,7 +21,7 @@ export const useDeleteAuditReport = () => {
   return useMutation({
     mutationFn: async (auditReportId: string) => {
       const { data } = await apiRequest.delete<TAuditReport>(
-        `/api/v1/audit-reports/${auditReportId}`
+        `/api/v1/insights/secrets/reports/${auditReportId}`
       );
       return data;
     },
