+++ frontend/src/hooks/api/auditReports/queries.tsx
@@ -30,7 +30,7 @@ export const useGetAuditReports = (
   return useQuery({
     queryKey: auditReportKeys.list(params),
     queryFn: async () => {
-      const { data } = await apiRequest.get<TGetAuditReportsResponse>("/api/v1/audit-reports", {
+      const { data } = await apiRequest.get<TGetAuditReportsResponse>("/api/v1/insights/secrets/reports", {
         params
       });
       return data;
